+++ app/src/main/java/org/simple/clinic/medicalhistory/newentry/NewMedicalHistoryUiActions.kt
@@ -6,7 +6,10 @@ interface NewMedicalHistoryUiActions {
   fun openPatientSummaryScreen(patientUuid: UUID)
   fun showOngoingHypertensionTreatmentErrorDialog()
   fun showDiagnosisRequiredErrorDialog()
+
+  fun showDiagnosisOrReferralRequiredErrorDialog()
   fun showHypertensionDiagnosisRequiredErrorDialog()
+  fun showHypertensionDiagnosisRequiredOrReferralErrorDialog()
   fun showChangeDiagnosisErrorDialog()
   fun showOngoingDiabetesTreatmentErrorDialog()