+++ app/src/main/java/org/simple/clinic/summary/PatientSummaryViewEffectHandler.kt
@@ -20,7 +20,10 @@ class PatientSummaryViewEffectHandler(
           currentFacility = viewEffect.currentFacility
       )
 
-      is ShowDiagnosisError -> uiActions.showDiagnosisError(viewEffect.diabetesManagementEnabled)
+      is ShowDiagnosisRequiredError -> uiActions.showDiagnosisRequiredError()
+      is ShowDiagnosisOrReferralRequiredError -> uiActions.showDiagnosisOrReferralRequiredError()
+      is ShowHypertensionDiagnosisRequiredError -> uiActions.showHypertensionDiagnosisRequiredError()
+      is ShowHypertensionDiagnosisOrReferralRequiredError -> uiActions.showHypertensionDiagnosisOrReferralRequiredError()
       is OpenContactPatientScreen -> uiActions.openPatientContactSheet(viewEffect.patientUuid)
       is NavigateToTeleconsultRecordScreen -> uiActions.navigateToTeleconsultRecordScreen(viewEffect.patientUuid, viewEffect.teleconsultRecordId)
       is OpenContactDoctorSheet -> uiActions.openContactDoctorSheet(viewEffect.patientUuid)